New: Daily reminder can now be set also for chores and option to show notifications   immediately (closes #149), also code improvements

+++ app/src/main/java/xyz/zedler/patrick/grocy/model/RecipePositionResolved.java
@@ -334,7 +334,7 @@ public class RecipePositionResolved {
     return calories;
   }
 
-  public void setCalories(int calories) {
+  public void setCalories(double calories) {
     this.calories = calories;
   }
